Added Announcement (#13)
* Release 2023.10.02

* Release 2023.10.01
- Refactor functions and classes to remove swift lint warnings and error
- Added platform parameter in session_start event

* Added privacy manifest file

* Refactor functions and classes (#11)

* Release 2023.10.01
- Refactor functions and classes to remove swift lint warnings and error
- Added platform parameter in session_start event

* Refactor

* SPM fix

* SPM fixed - 2

* Announcement support

* Fixed UI issues

* Announcement UI issue fixed

* Announcement issue fixed, Added push notification

* Announcement issue fixed, update push token if not generate

* Banner issue fixed

* SPM fixes

* announcement with dev for SPM

* SPM fix0 resource path to quill html

* SPM fixes

* Announcement

* Obj c fixes

* fixed progress bar issue

+++ Survey/Classes/Survey Manager/View Controllers/OFRatingViewController+Display.swift
@@ -36,7 +36,7 @@ extension OFRatingViewController {
                             }
                         let extraScreens = (surveyScreens.count - filteredSurveyScreens.count) > 0 ? 1 : 0
                         self.progressBar.setProgress(
-                            Float((self.currentScreenIndex + 1)/(filteredSurveyScreens.count + extraScreens)),
+                            Float(CGFloat(self.currentScreenIndex + 1)/CGFloat(filteredSurveyScreens.count + extraScreens)),
                             animated: true
                         )
                     }
